refactor: arborescence service messaging
