+++ backend/lib/messaging/email/email-sending-tool.ts
@@ -1,9 +1,9 @@
 import { ArgumentParser } from "argparse"
-import config from "../config/index.js"
-import { EmailCategory } from "../../lib/enums/email.js"
+import config from "../../../config/index.js"
+import { EmailCategory } from "../../../../lib/enums/email.js"
 import mongoose from "mongoose"
-import mongooseConfig from "../config/mongoose.js"
-import { processSendEmails } from "./emails/sending.js"
+import mongooseConfig from "../../../config/mongoose.js"
+import { processSendEmails } from "../sending.js"
 
 mongooseConfig(mongoose, config)
